requirements.txt
openai>=1.0.0
tavily-python
python-dotenv
requests
beautifulsoup4
pydantic
beautifulsoup4
pytest